feat: update next-page-dir boilerplate
