+++ boilerplate/fullstack/next/README.md
@@ -1,0 +1,74 @@
+# SuperTokens + Next
+
+A demo implementation of [SuperTokens](https://supertokens.com/) with [Next.js](https://nextjs.org/). Using the page router.
+
+## General Info
+
+This project aims to demonstrate how to integrate SuperTokens into a Nexjs application. Its primary purpose is to serve as an educational tool, but it can also be used as a starting point for your own project.
+
+### Config
+
+#### SuperTokens
+
+The full configuration needed for SuperTokens (the frontend part) to work is in the `src/config.tsx` file. This file will differ based on the [auth recipe](https://supertokens.com/docs/guides) you choose.
+
+If you choose to use this as a starting point for your own project, you can further customize the options and config in the `src/config.tsx` file. Refer to our [docs](https://supertokens.com/docs) (and make sure to choose the correct recipe) for more details.
+
+## Application Flow
+
+The application consists of four main parts:
+
+1. **Root Component (`_app.tsx`)**
+
+    - Initializes SuperTokens with the provided configuration (generated by the CLI)
+    - Wraps the application with necessary providers:
+        - `SuperTokensWrapper`: Manages auth state and session
+        - `ComponentWrapper`: Provides auth UI customization for specific auth recipes. For example, if you choose to use the `ThirdPartyPasswordless` recipe, the `ComponentWrapper` will provide the UI customization for the passwordless login flow (showing a disclaimer about the SMS delivery in demo apps).
+
+2. **Home Component (`/` route, `/index.tsx` component)**
+
+    - Public landing page accessible to all users
+    - Provides navigation to authentication and dashboard
+    - Displays basic application information and links
+
+3. **Dashboard Component (`/dashboard` route, `/Dashboard/index.tsx` component)**
+    - Protected route only accessible to authenticated users
+    - Protected by `SessionAuth` component
+    - Displays user information and session details
+    - Provides functionality to:
+        - View user ID
+        - Call test API endpoints
+        - Access documentation
+        - Sign out
+
+When a user visits the application, they start at the home page (`/`). They can choose to authenticate through the `/auth` route, and once authenticated, they gain access to the protected dashboard. The session state is managed throughout the application using SuperTokens' session management.
+
+## Customizations
+
+If you want to customize the default auth UI, you have two options:
+
+1. Refer to the [docs](https://supertokens.com/docs/thirdpartyemailpassword/advanced-customizations/react-component-override/usage) on how to customize the pre-built UI.
+2. Roll your own UI by choosing "Custom UI" in the right sidebar in the [docs](https://supertokens.com/docs/thirdpartyemailpassword/quickstart/frontend-setup).
+
+## Additional resources
+
+-   Custom UI Example: https://github.com/supertokens/supertokens-web-js/tree/master/examples/react/with-thirdpartyemailpassword
+-   Custom UI Blog post: https://supertokens.medium.com/adding-social-login-to-your-website-with-supertokens-custom-ui-only-5fa4d7ab6402
+-   Awesome SuperTokens: https://github.com/kohasummons/awesome-supertokens
+
+## Contributing
+
+Please refer to the [CONTRIBUTING.md](https://github.com/supertokens/create-supertokens-app/blob/master/CONTRIBUTING.md) file in the root of the [`create-supertokens-app`](https://github.com/supertokens/create-supertokens-app) repo.
+
+## Contact us
+
+For any questions, or support requests, please email us at team@supertokens.io, or join our [Discord](https://supertokens.io/discord) server.
+
+## Notes
+
+-   To know more about how this app works and to learn how to customise it based on your use cases refer to the [SuperTokens Documentation](https://supertokens.com/docs/guides)
+-   We have provided development OAuth keys for the various built-in third party providers in the `/app/config/backend.ts` file. Feel free to use them for development purposes, but **please create your own keys for production use**.
+
+## Authors
+
+Created with :heart: by the folks at SuperTokens.io.
