+++ boilerplate/fullstack/next/config/appInfo.ts
@@ -5,7 +5,7 @@ const apiBasePath = "/api/auth/";
 export const websiteDomain = process.env.APP_URL || process.env.NEXT_PUBLIC_APP_URL || `http://localhost:${port}`;
 
 export const appInfo = {
-    appName: "SuperTokens Demo App",
+    appName: "SuperTokens + Next.js",
     websiteDomain,
     apiDomain: websiteDomain,
     apiBasePath,
