+++ boilerplate/fullstack/next/config/frontendConfig.tsx
@@ -6,6 +6,11 @@ export let frontendConfig = () => {
         // recipeList contains all the modules that you want to
         // use from SuperTokens. See the full list here: https://supertokens.com/docs/guides
         recipeList: [],
+        getRedirectionURL: async (context) => {
+            if (context.action === "SUCCESS" && context.newSessionCreated) {
+                return "/dashboard";
+            }
+        },
     };
 };
